Advanced button system support progress.
Updated the way the sketch api reference object is obtained.
Added support for a Label Placeholder to take the place of the Label layer.
Removed the requirement of the Background layer.
Added utilities for getting the master symbol for a layer, artboard for a layer, and getting a layer by name inside of layer group (or itself).
Added support for logging to the Sketch "run script" console instead of the Safari console, when the "run script" console is available.
Wrote most of the documentation on the new advanced button system.